Massive refactoring of tests to allow for testing multiple versions of react router
