+++ package.json
@@ -13,9 +13,9 @@
     "build": "mkdir -p lib && babel ./src --out-dir ./lib",
     "lint": "eslint src test",
     "prepublish": "rm -rf lib && npm run build",
-    "test": "mocha --compilers js:babel-core/register --recursive --require test/init.js test/**/*-test.js",
-    "test:cov": "babel-node --max-old-space-size=4076 $(npm bin)/babel-istanbul cover $(npm bin)/_mocha -- --require test/init.js test/**/*-test.js",
-    "test:watch": "mocha --compilers js:babel-core/register --recursive --require test/init.js -w test/**/*-test.js"
+    "test": "mocha --compilers js:babel-core/register --recursive --require test/init.js test/rrv3-test.js",
+    "test:cov": "babel-node --max-old-space-size=4076 $(npm bin)/babel-istanbul cover $(npm bin)/_mocha -- --require test/init.js test/rrv3-test.js",
+    "test:watch": "mocha --compilers js:babel-core/register --recursive --require test/init.js -w test/rrv3-test.js"
   },
   "repository": {
     "type": "git",
